make utils a package
