+++ src/pie_modules/utils/__init__.py
@@ -1,0 +1,2 @@
+from .dictionary import flatten_dict, list_of_dicts2dict_of_lists
+from .hydra import resolve_type
